[General] Delete obsolete mcp_input_parser
That file was very old and cause more confusion than anything. So, the functions
have been moved to the appropriate data_handler file, which contains
all the functions that we use to handle the external data.
Also, some function names were rewritten in camel case so that they look more
in line with the other function names.